EJOIN: Rename match_xyz file to match
